refactor: Refact store providers for page modules
