+++ src/core/containers/Base/Detail/Page/index.jsx
@@ -17,10 +17,8 @@
  */
 
 import React from 'react'
-import { pick, isEmpty } from 'lodash'
 import { NavLink, withRouter } from 'react-router-dom'
 import { inject, Provider } from 'mobx-react'
-import { Loading } from '@pitrix/lego-ui'
 import pathToRegexp from 'path-to-regexp'
 
 import { ICON_TYPES } from 'utils/constants'
@@ -33,39 +31,7 @@ import styles from './index.scss'
 class DetailPage extends React.Component {
   constructor(props) {
     super(props)
-    this.state = { initializing: true }
     this.stores = {}
-  }
-
-  componentDidMount() {
-    this.init()
-  }
-
-  async init() {
-    const params = pick(this.props.match.params, [
-      'cluster',
-      'workspace',
-      'namespace',
-      'project_id',
-    ])
-
-    if (params.workspace) {
-      await this.props.rootStore.getRules({
-        workspace: params.workspace,
-      })
-    } else if (params.cluster) {
-      await this.props.rootStore.getRules({
-        cluster: params.cluster,
-      })
-    }
-
-    if (!isEmpty(params)) {
-      await this.props.rootStore.getRules({
-        ...params,
-        project: params.namespace || params.project_id,
-      })
-    }
-    this.setState({ initializing: false })
   }
 
   get authKey() {
@@ -115,10 +81,6 @@ class DetailPage extends React.Component {
 
   render() {
     const { routes, stores, ...sideProps } = this.props
-
-    if (this.state.initializing) {
-      return <Loading className="ks-page-loading" />
-    }
 
     return (
       <Provider {...this.stores} {...stores}>
