--- src/core/layouts/List/index.scss
@@ -1,23 +1,0 @@
-@import '~scss/variables';
-
-.main {
-  overflow-y: auto;
-  margin-left: $nav-width - 20px;
-  padding: 20px;
-}
-
-.navWrapper {
-  position: fixed;
-  top: 88px;
-  width: $nav-width;
-  height: calc(100vh - 108px);
-  padding: 0 20px 40px 20px;
-  overflow-y: auto;
-}
-
-.nav {
-  width: auto;
-  padding: 0;
-  z-index: 1;
-}
-
