--- src/pages/clusters/containers/layout.scss
@@ -1,86 +1,0 @@
-@import '~scss/variables';
-
-.wrapper {
-  max-width: 1400px;
-  padding: 0 20px;
-  padding-top: 40px;
-  margin: 0 auto;
-
-  p {
-    color: #79879c;
-  }
-}
-
-.main {
-  overflow-y: auto;
-  margin-left: $nav-width - 20px;
-  padding: 20px;
-}
-
-.navWrapper {
-  position: fixed;
-  top: 88px;
-  width: $nav-width;
-  height: calc(100vh - 108px);
-  padding: 0 20px 40px 20px;
-  overflow-y: auto;
-}
-
-.nav {
-  width: auto;
-  padding: 0;
-  z-index: 1;
-}
-
-.cards {
-  margin-top: 32px;
-  padding-bottom: 32px;
-  list-style: none;
-
-  :global .h6 {
-    line-height: 1.43;
-    margin-bottom: 20px;
-  }
-}
-
-.title {
-  margin-bottom: 32px;
-}
-
-.create {
-  width: 96px;
-}
-
-.total {
-  display: inline-block;
-  height: 20px;
-  padding: 0 8px;
-  margin-left: 8px;
-  border-radius: 10px;
-  background-color: #abb4be;
-  font-weight: 500;
-  color: #ffffff;
-}
-
-
-.noData {
-  height: 226px;
-  border-radius: $border-radius;
-  padding: 32px 0;
-  box-shadow: 0 4px 8px 0 rgba(36, 46, 66, 0.06);
-  background-color: #ffffff;
-  text-align: center;
-
-  p {
-    width: 586px;
-    margin: 0 auto;
-    margin-top: 20px;
-  }
-}
-
-.loading {
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-}
