+++ src/pages/projects/containers/ConfigMaps/Detail/index.jsx
@@ -54,9 +54,12 @@ export default class ConfigMapDetail extends React.Component {
 
   get listUrl() {
     const { cluster, workspace, namespace } = this.props.match.params
-    return `${
-      workspace ? `/${workspace}` : ''
-    }/clusters/${cluster}/projects/${namespace}/${this.module}`
+    if (workspace) {
+      return `/${workspace}/clusters/${cluster}/projects/${namespace}/${
+        this.module
+      }`
+    }
+    return `/clusters/${cluster}/${this.module}`
   }
 
   fetchData = () => {
