+++ src/pages/projects/containers/DaemonSets/Detail/index.jsx
@@ -53,10 +53,13 @@ export default class DaemonSetDetail extends React.Component {
   }
 
   get listUrl() {
-    const { cluster, workspace, namespace } = this.props.match.params
-    return `${
-      workspace ? `/${workspace}` : ''
-    }/clusters/${cluster}/projects/${namespace}/${this.module}`
+    const { workspace, cluster, namespace } = this.props.match.params
+    if (workspace) {
+      return `/${workspace}/clusters/${cluster}/projects/${namespace}/${
+        this.module
+      }`
+    }
+    return `/clusters/${cluster}/${this.module}`
   }
 
   fetchData = () => {
