+++ src/pages/projects/containers/Deployments/Detail/ResourceStatus/index.jsx
@@ -237,7 +237,5 @@ class ResourceStatus extends React.Component {
   }
 }
 
-export default inject('detailStore', 'projectStore', 'fedDetailStore')(
-  observer(ResourceStatus)
-)
+export default inject('detailStore', 'fedDetailStore')(observer(ResourceStatus))
 export const Component = ResourceStatus
