+++ src/pages/projects/containers/Deployments/Detail/index.jsx
@@ -53,9 +53,12 @@ export default class DeploymentDetail extends React.Component {
 
   get listUrl() {
     const { workspace, cluster, namespace } = this.props.match.params
-    return `${
-      workspace ? `/${workspace}` : ''
-    }/clusters/${cluster}/projects/${namespace}/${this.module}`
+    if (workspace) {
+      return `/${workspace}/clusters/${cluster}/projects/${namespace}/${
+        this.module
+      }`
+    }
+    return `/clusters/${cluster}/${this.module}`
   }
 
   get routing() {
