+++ src/pages/projects/containers/Jobs/Detail/index.jsx
@@ -60,9 +60,12 @@ export default class JobDetail extends React.Component {
 
   get listUrl() {
     const { workspace, cluster, namespace } = this.props.match.params
-    return `${
-      workspace ? `/${workspace}` : ''
-    }/clusters/${cluster}/projects/${namespace}/${this.module}`
+    if (workspace) {
+      return `/${workspace}/clusters/${cluster}/projects/${namespace}/${
+        this.module
+      }`
+    }
+    return `/clusters/${cluster}/${this.module}`
   }
 
   fetchData = () => {
