+++ src/pages/workspaces/App.jsx
@@ -16,11 +16,66 @@
  * along with KubeSphere Console.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { inject } from 'mobx-react'
+import React, { Component } from 'react'
+import { inject, observer, Provider } from 'mobx-react'
+import { Loading } from '@pitrix/lego-ui'
+
 import { renderRoutes } from 'utils/router.config'
+
+import WorkspaceStore from 'stores/workspace'
 
 import routes from './routes'
 
-const App = () => renderRoutes(routes)
+@inject('rootStore')
+@observer
+class WorkspaceLayout extends Component {
+  constructor(props) {
+    super(props)
 
-export default inject('rootStore')(App)
+    this.store = new WorkspaceStore()
+
+    this.init(props.match.params)
+  }
+
+  componentDidUpdate(prevProps) {
+    if (prevProps.match.params.workspace !== this.workspace) {
+      this.init(this.props.match.params)
+    }
+  }
+
+  async init(params) {
+    this.store.initializing = true
+
+    await Promise.all([
+      this.store.fetchDetail(params),
+      this.store.fetchClusters({ ...params, limit: -1 }),
+      this.props.rootStore.getRules(params),
+    ])
+
+    globals.app.cacheHistory(this.props.match.url, {
+      type: 'Workspace',
+      name: this.store.detail.name,
+      description: this.store.detail.description,
+    })
+
+    this.store.initializing = false
+  }
+
+  get workspace() {
+    return this.props.match.params.workspace
+  }
+
+  render() {
+    const { initializing } = this.store
+
+    if (initializing) {
+      return <Loading className="ks-page-loading" />
+    }
+
+    return (
+      <Provider workspaceStore={this.store}>{renderRoutes(routes)}</Provider>
+    )
+  }
+}
+
+export default WorkspaceLayout
